Complete Overhaul to Windows Forms
This is a complete overhaul to give this program a GUI.

I reworked the code to function on a WindowsForms program as opposed to a console-based application. This makes the program much more user friendly.
In addition, I also implemented other features to make the program more user-friendly. There is now a button to contact the developer (me) that opens my twitter page, and there is a button to enable/disable debug mode during runtime.